moved the components to the components folder
