+++ components/RaffleTicket.tsx
